refactor #37 : 재고 반영 시점 변경
- 결제 프로세스 도입 시 재고 감소
- 20%확률로 고객이 취소 -> 재고 증가
- 결제 시도 시 20%확률로 고객귀책 사유 실패 -> 재고 증가
- 수량 부족 시 결제 실패

+++ module_order/src/main/java/com/hanghae/module_order/client/dto/response/ItemDetailsResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
@@ -15,7 +17,7 @@ public class ItemDetailsResponse {
     private Long sellerNum;
     private String title;
     private String description;
-    private Long price;
+    private BigDecimal price;
     private Long stock;
     private LocalDateTime availableAt;
     private LocalDateTime endAt;
